# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For JPO ERP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/token/c/{jumlah}": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Create a new token",
                "operationId": "createToken",
                "parameters": [
                    {
                        "name": "jumlah",
                        "in": "path",
                        "description": "Input total token to create",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "Token to generate other tokens",
                                        "type": "string",
                                        "example": "adibkeren"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success create new Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "success create token"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/tokens/": {
            "get": {
                "tags": [
                    "Token"
                ],
                "summary": "Get all Tokens",
                "description": "Get All Tokens",
                "operationId": "indexTokens",
                "responses": {
                    "200": {
                        "description": "success fetch Tokens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Tokens": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failed to fetch Tokens.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/token/{id}": {
            "get": {
                "tags": [
                    "Token"
                ],
                "summary": "Get token by id.",
                "description": "Get Token by id.",
                "operationId": "showToken",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/token/d/{id}": {
            "delete": {
                "tags": [
                    "Token"
                ],
                "summary": "Delete token by ID",
                "description": "Delete Token by ID.",
                "operationId": "deleteToken",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the token to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failed to delete Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}